requirements.txt
requests
pandas
google-cloud-bigquery
apache-airflow-providers-google